[Refactor] message input change uncontrolled component#54
