+++ pong-game/components/Chat/ChatInfoLog/ChatPassword.module.scss
@@ -32,32 +32,32 @@
     height: 30px;
     padding: 8px 16px;
     margin-top: 32px;
-    .passwordInput {
-      width: 100%;
-      height: 100%;
-      font-size: 1rem;
-      border: none;
-      border-radius: 10px;
-      padding: 8px 16px;
-      box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-      &:focus {
-        outline: none;
-      }
-      &::placeholder {
-        font-size: 1rem;
-        text-align: center;
-      }
-    }
-    .passwordInputPlaceHolder {
-      position: absolute;
-      padding: 8px;
-      pointer-events: none;
-      font-size: 1rem;
-      transition: 0.5s;
-      color: $gray_scale_400;
-      left: 15px;
-      font-family: 'Pretendard-SemiBold';
-    }
+    // .passwordInput {
+    //   width: 100%;
+    //   height: 100%;
+    //   font-size: 1rem;
+    //   border: none;
+    //   border-radius: 10px;
+    //   padding: 8px 16px;
+    //   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    //   &:focus {
+    //     outline: none;
+    //   }
+    //   &::placeholder {
+    //     font-size: 1rem;
+    //     text-align: center;
+    //   }
+    // }
+    // .passwordInputPlaceHolder {
+    //   position: absolute;
+    //   padding: 8px;
+    //   pointer-events: none;
+    //   font-size: 1rem;
+    //   transition: 0.5s;
+    //   color: $gray_scale_400;
+    //   left: 15px;
+    //   font-family: 'Pretendard-SemiBold';
+    // }
     .submitPassword {
       background-color: transparent;
       position: absolute;
@@ -69,11 +69,11 @@
   }
 }
 
-.passwordInput:valid ~ .passwordInputPlaceHolder,
-.passwordInput:focus ~ .passwordInputPlaceHolder {
-  transform: translateY(-2.2rem);
-  padding: 0 16px;
-  background-color: transparent;
-  font-family: 'Pretendard-SemiBold';
-  color: black;
-}
+// .passwordInput:valid ~ .passwordInputPlaceHolder,
+// .passwordInput:focus ~ .passwordInputPlaceHolder {
+//   transform: translateY(-2.2rem);
+//   padding: 0 16px;
+//   background-color: transparent;
+//   font-family: 'Pretendard-SemiBold';
+//   color: black;
+// }
