+++ pong-game/components/Chat/Input/MessageInput.module.scss
@@ -1,0 +1,37 @@
+.messageInput {
+  width: 100%;
+  height: 100%;
+  outline: none;
+  border: none;
+  background-color: #fff;
+  border-radius: 10px;
+  font-size: 1rem;
+  padding: 0px 16px;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  &:focus {
+    outline: none;
+  }
+  &::placeholder {
+    font-size: 1rem;
+    text-align: center;
+  }
+}
+.folderNameInputPlaceHolder {
+  position: absolute;
+  padding: 10px;
+  pointer-events: none;
+  font-size: 1rem;
+  text-transform: uppercase;
+  transition: 0.5s;
+  color: #c0bebe;
+  font-family: 'Pretendard-SemiBold';
+}
+
+.messageInput:valid ~ .folderNameInputPlaceHolder,
+.messageInput:focus ~ .folderNameInputPlaceHolder {
+  transform: translateY(-1.7rem);
+  padding: 0 16px;
+  background-color: transparent;
+  color: black;
+  font-family: 'Pretendard-SemiBold';
+}
