+++ pong-game/components/Chat/Input/MessageInput.tsx
@@ -1,0 +1,17 @@
+import { RefObject } from 'react'
+import styles from './MessageInput.module.scss'
+
+interface MessageInputProps {
+  messageRef: RefObject<HTMLInputElement>
+}
+
+function MessageInput(props: MessageInputProps): JSX.Element {
+  return (
+    <>
+      <input type="text" className={styles.messageInput} ref={props.messageRef} required />
+      <span className={styles.folderNameInputPlaceHolder}>Message Input</span>
+    </>
+  )
+}
+
+export default MessageInput
